### 2021-02-25 #3064 @whedon
[PRE REVIEW]: PyPortfolioOpt: portfolio optimization in Python
**Submitting author:** @robertmartin8 (<a href="http://orcid.org/0000-0002-5225-7700">Robert Martin</a>)
**Repository:** <a href="https://github.com/robertmartin8/PyPortfolioOpt/" target ="_blank">https://github.com/robertmartin8/PyPortfolioOpt/</a>
**Version:** v1.4.1
