### 2021-02-25 #3064 @kthyng
@robertmartin8 looks like you have some errors in your paper ( see above). You can check the compilation of your paper after making changes by running the same whedon command I did or you can use the methods described [here](https://joss.readthedocs.io/en/latest/submitting.html#checking-that-your-paper-compiles).
