### 2021-02-26 #3064 @robertmartin8
@kthyng Strange, it compiled without issues on docker and the online service. The compiled paper from `whedon.theoj.org` is attached:

[suwa48ol9jb6h7tzgxyi.pdf](https://github.com/openjournals/joss-reviews/files/6046927/suwa48ol9jb6h7tzgxyi.pdf)
