### 2021-02-26 #3064 @vissarion
@robertmartin8 thanks for resolving the issue on generated pdf. We should be ready start the review now.
