### 2021-04-12 robertmartin8/PyPortfolioOpt#319 @omendezmorales
typo in Quick example section?
**What are you trying to do?**
In the [quick example](https://github.com/robertmartin8/PyPortfolioOpt#a-quick-example) section, calling
`da = DiscreteAllocation(cleaned_weights, latest_prices, total_portfolio_value=10000) `
 retrieves
"NameError: name 'weights' is not defined"
**What have you tried?**
I believe given the variable used previously, it should be `cleaned_weights`?

**What data are you using?**
I am using the following dataset ("tests/resources/stock_prices.csv") as provided in the repo, and executing the sample code in a _ipython_ environment

Thanks for your help.
