### 2021-04-12 091df3f
Fix quick example in README
