### 2021-04-12 robertmartin8/PyPortfolioOpt#319 @phschiele
Good catch, indeed, using `cleaned_weights` works.
@robertmartin8, I have changed it here: #320
