### 2021-04-13 robertmartin8/PyPortfolioOpt#319 @robertmartin8
Cheers, fixed!
