### 2021-04-13 #3066 @robertmartin8
Hi @omendezmorales,

Really appreciate you taking the time to go through the package and to provide thoughtful feedback! I've addressed the documentation issue you raised – good spot.

The `GLPK_MI` issue is quite a thorny one. `GLPK_MI` is supposed to come installed with CVXOPT, so in the past, a quick `pip install cvxopt` did the job. More recently, I've had some issues with that (as per the thread you linked). I'd love to use one of CVXPY's defaults, but unfortunately they no longer recommend the use of `ECOS_BB` to solve mixed-integer problems due to correctness issues. When and if they ship a default mixed-integer problem, all these problems will go away.

In the meantime, I've changed the documentation as you suggested so that new users don't get discouraged by a packaging issue. The good news is that `lp_portfolio` and `greedy_portfolio` give very similar answers in most cases, so user experience shouldn't be impacted too much.

Thanks again for the comments and let me know if you've got any other feedback or suggestions!

Cheers,
Robert
