### 2021-04-22 #3066 @vissarion
:wave: @Gunratan, @daviddewhurst, @AvianaGlobal, @syoh would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
