### 2021-04-28 #3066 @SteveDiamond
> Hi @robertmartin8,
> I'm contacting you in the context of the review of your article for the [JOSS Journal](https://github.com/openjournals/joss-reviews/issues/3066).
>
> I want to congratulate you for the effort you have put in PyPortfolioOpt, and just draw your attention to the simple documentation [issue I filed](https://github.com/robertmartin8/PyPortfolioOpt/issues/319), which deserves (maybe) a modification.
>
> Another possible improvement related to this _quick example_: I see the default solver in the call to `lp_portfolio()` is GLPK_MI. I've had difficulty installing it, and see [some others have too](https://github.com/cvxgrp/cvxpy/issues/972) (I'm using an Ubuntu 20.04 distribution to try things out). Given this example might be the _first impression_ for newcomers, you may consider either changing the default solver used (e.g. using cvxpy's default installed solvers ), or adapt the example instead. Otherwise, newcomers will see
>
> ```
> NameError                                 Traceback (most recent call last)
> <ipython-input-13-1838607f55a8> in <module>
> ----> 1 allocation, leftover = da.lp_portfolio()
>
> /usr/local/lib/python3.8/dist-packages/pypfopt/discrete_allocation.py in lp_portfolio(self, reinvest, verbose, solver)
>     322
>     323         if solver is not None and solver not in cp.installed_solvers():
> --> 324             raise NameError("Solver {} is not installed. ".format(solver))
>     325         opt.solve(solver=solver)
>     326
>
> NameError: Solver GLPK_MI is not installed.
> ```
>
> Once more, kudos for your work, and await to read your comments.

@robertmartin8 It should be fine if you use the ECOS_BB solver. We dropped it as the default MIP solver, but you can still access it via ``problem.solve(solver="ECOS_BB")``.
