### 2021-04-28 #3066 @robertmartin8
Hey @SteveDiamond, cheers for doing this!

Noted re `ECOS_BB` – I should probably add unit tests to check if it agrees with `GLPK_MI`.
