### 2021-05-04 #3066 @vissarion
@SteveDiamond indeed I invited you to be a reviewer. I though I sent you the instructions but maybe I didn't. So, please find below a few useful links
https://joss.readthedocs.io/en/latest/review_criteria.html
https://joss.readthedocs.io/en/latest/review_checklist.html
https://joss.readthedocs.io/en/latest/reviewer_guidelines.html

Your checklist is on top of this thread. Please let me know if you have any concerns or questions.

Thank you very much again for accepting to review this JOSS submission!
