### 2021-05-05 robertmartin8/PyPortfolioOpt#331 @SteveDiamond
Minor documentation issues
**Describe the bug**
I found the following minor errors in the documentation:

In the User Guide:
```
ef = EfficientFrontier(mu, S)
ef.add_objective(objective_functions.L2_reg, gamma=0.1)
w = ef.max_sharpe()
print(ef.clean_weights())
```
should start with ``from pypfopt import objective_functions``.

In the FAQ "or small problems with ..." should be "For small problems with ...".
