### 2021-05-05 #3066 @SteveDiamond
I looked over everything, and it's good to go from my perspective. The documentation is very thorough. I found a few typos (https://github.com/robertmartin8/PyPortfolioOpt/issues/331) but no major issues.
