### 2021-05-05 robertmartin8/PyPortfolioOpt#331 @robertmartin8
Cheers, fixed!
