### 2021-05-06 #3066 @robertmartin8
@vissarion

I made a [tagged release](https://github.com/robertmartin8/PyPortfolioOpt/releases/tag/v1.4.1)  (v1.4.1) and the zenodo DOI is 10.5281/zenodo.4740186.
