### 2021-05-06 #3066 @vissarion
> @vissarion
>
> I made a [tagged release](https://github.com/robertmartin8/PyPortfolioOpt/releases/tag/v1.4.1) (v1.4.1) and the zenodo DOI is [10.5281/zenodo.4740186](https://doi.org/10.5281/zenodo.4740186).

thanks. You have to make sure that the list of authors is the same in the paper and zenodo.
