### 2021-05-06 #3066 @kthyng
Hi @robertmartin8! Please change the metadata in your Zenodo archive so that the title and author list exactly match your JOSS paper.
