### 2021-05-06 #3066 @kthyng
Hi @robertmartin8! Your paper looks great, except the references aren't all properly capitalized. In a quick glance, I saw "python" not capitalized" and "i" not capitalized. Please go through your references in detail and add {} around characters, words, or phrases in which you want to preserve the capitalization. You can re-compile your paper after you've made changes to make sure all the references look good at that point, then ping back here and we can wrap it up!
