### 2021-05-07 #3066 @robertmartin8
@kthyng Cheers for the LaTeX tip – didn't know about the curly braces thing!

I've fixed the obvious error and have tried to match the capitalisation of paper titles to that of the actual published articles (though in some cases there is ambiguity and I left it as lower case).
